liabilities-calc-script/src/requirements.txt
polars-lts-cpu==1.23.0